# tsdb: WAL replay waitUntilIdle() optimize sleep time