@roidelapluie 2022-06-17
I wonder if we can the get back to the old, safer value of 1ms?