@bboreham 2022-06-17
> if we need to

Citation needed.

I.e. how do you get from that notion to the code change proposed?

Can you write a test that clearly sets up the “need to” case, and fails if you do not have this code?