@colega 2022-06-30
I think it is _valid_ but it's not the optimal solution. 

The logic before was: "Throw away at most 1 reusable buffer every 10 ms until queue is empty. Wait 10ms between checks"

Now the logic is: "Throw away every single buffer, and re-create the 10ms timer. Check the length when one of both things (throwing a buffer or 10ms) happens."

I don't think (assumption alert!) that processing a bunch of samples takes more than 10ms so this now looks just like a fix to the fact that everything is non-deterministic in this logic (Why do we drain the output in the previous lines instead of just not pushing into the output when it's full? Why don't we properly signal that we're done? Why can't we rely on an output arriving to check the length?). I can understand that benchmarks are faster (now we don't have to wait those 10ms whenever the for loop condition happens), but also we're throwing away reusable buffers which has a negative effect.

I talked about this to @jesusvazquez offline and [proposed an alternative approach](https://github.com/colega/prometheus/commit/8490579d695a533157c1acd394dc368eaf0b7a81) that would be fully deterministic and would have some extra optimizations (we don't need to unlock/lock between slices of incoming samples, which I guess is the happy case when processing WAL).

I guess @jesusvazquez has been to occupied on the OOO task and he forgot to update this.