@jesusvazquez 2022-07-03
Hello!

Yes sorry that I did not have time to address the feedback here yet. I do have some local work attempting to remove the lock mechanism that I was investigating with @codesome (meaning we would only keep the waiting) and also optimizing the waiting to only wait if a series was not created (if it was creates means its first time we see it so no need to wait)

I've been a little busy this past couple of weeks with out-of-order, I'll try to submit these changes next week.