@bboreham 2022-07-04
Thanks @colega. If I can rephrase your explanation:

* We push a blank input then watch for the input queue to reach zero length.
* No-one else is pushing to the input, so once the one we pushed has been consumed we know the worker is doing no real work, which is equivalent to “idle”.
* Each time the worker finishes one item it pushes to the output chan, so if we see that happen that means it might have just started on our blank input; we should check the input queue again.

This seems valid, thank you. I'm happy now that this PR was merged.

> we're throwing away reusable buffers which has a negative effect

Yes, we will drain the entire output channel, and therefore need to re-allocate buffers when we start the next record.  But from the benchmark that seems to come out ahead.

> Why do we drain the output in the previous lines instead of just not pushing into the output when it's full?

I didn't think of that. It's worth trying.

> Why can't we rely on an output arriving to check the length?

Suppose everything was idle when we enter `waitForIdle`.  Then waiting for output means we wait for the goroutine to finish the blank input, rather than waiting for it to start. It's not _wrong_, but doing a bit more than what it says.
(I'm being very pedantic over words. When trying to avoid race conditions it pays to be extremely careful.)

> Why don't we properly signal that we're done? 

Careful - we don't want to know it is "done one piece of work" (which we can tell from the lock) - we want to know it has "done all queued work". But I agree it's worth trying to find a clean way to do that.