@codesome 2022-07-04
> > Careful - we don't want to know it is "done one piece of work" (which we can tell from the lock) - we want to know it has "done all queued work". But I agree it's worth trying to find a clean way to do that.
> 
> I saw that, and [colega@8490579](https://github.com/colega/prometheus/commit/8490579d695a533157c1acd394dc368eaf0b7a81) should handle that properly.

Interesting approach, but the idea is to not pause the processing, rather to empty the input channel (after processing i.e.), because there might be some records in the input channel that needs to be un-done when we see a duplicate series record. The name `waitUntilIdle()` is confusing, it is technically `waitUntilEmpty()`.