@colega 2022-07-04
But in fact it's pausing the processing (because the caller won't send any other samples until that task is done)