@codesome 2022-07-04
True. That is a side effect of emptying the input buffer because WAL replay happens in a serial fashion, which means you cannot process the new sample records before processing the current series record.